feat(ui5-sidenavigation): add base class
