+++ packages/fiori/src/SideNavigation.hbs
@@ -12,7 +12,8 @@
 				   aria-haspopup="{{_ariaHasPopup}}"
 				   aria-checked="{{selected}}"
 				   title="{{_tooltip}}"
-				   href="#page1"
+				   href="{{_href}}"
+				   target="{{_target}}"
 				>
 					<ui5-icon class="ui5-sn-item-icon" name="{{icon}}"></ui5-icon>
 					<div class="ui5-sn-item-text">{{text}}</div>
@@ -31,7 +32,8 @@
 					   aria-current="{{_ariaCurrent}}"
 					   title="{{_tooltip}}"
 					   aria-owns="{{_groupId}}"
-					   href="#page1"
+					   href="{{_href}}"
+					   target="{{_target}}"
 					>
 						<ui5-icon class="ui5-sn-item-icon" name="{{icon}}"></ui5-icon>
 						<div class="ui5-sn-item-text">{{text}}</div>
@@ -56,7 +58,8 @@
 									   data-sap-focus-ref
 									   tabindex="{{_tabIndex}}"
 									   aria-current="{{_ariaCurrent}}"
-									   href="#page1"
+									   href="{{_href}}"
+									   target="{{_target}}"
 									>
 										<ui5-icon class="ui5-sn-item-icon" name="{{icon}}"></ui5-icon>
 										<div class="ui5-sn-item-text">{{text}}</div>
