+++ packages/fiori/test/pages/SideNavigationOnly.html
@@ -17,7 +17,7 @@
 		</ui5-side-navigation-item>
 		<ui5-side-navigation-item text="Locations" icon="locate-me" selected></ui5-side-navigation-item>
 		<ui5-side-navigation-item expanded text="Events.............................................................." icon="calendar" whole-item-toggleable>
-			<ui5-side-navigation-sub-item text="Local.............................................................."></ui5-side-navigation-sub-item>
+			<ui5-side-navigation-sub-item href="#page1" target="_self" text="Local.............................................................."></ui5-side-navigation-sub-item>
 			<ui5-side-navigation-sub-item text="Others"></ui5-side-navigation-sub-item>
 		</ui5-side-navigation-item>
